+++ examples/unicorn_depthcache_cluster_for_binance/ubdcc-get-cluster-info.py
@@ -5,32 +5,26 @@
 import asyncio
 import logging
 import os
+from dotenv import load_dotenv
 from pprint import pprint
 from unicorn_binance_local_depth_cache import BinanceLocalDepthCacheManager, DepthCacheClusterNotReachableError
 
-footer: str = "By LUCIT - www.lucit.tech"
+load_dotenv()
+
 exchange: str = "binance.com-futures"
-limit_count: int = 2
 markets: list = ['1000SHIBUSDT', 'BTCUSDT', 'ETHUSDT']
-title: str = "UBLDC Demo"
-threshold_volume: float = 200000.0
-threshold_volume_limit_count: int = 3
-update_interval_ms: int = 100
-
-api_secret = ""
-license_token = ""
-ubdcc_address = ""
-
+ubdcc_address: str = os.getenv('UBDCC_ADDRESS')
 
 logging.getLogger("unicorn_binance_local_depth_cache")
-logging.basicConfig(level=logging.INFO,
+logging.basicConfig(level=logging.ERROR,
                     filename=os.path.basename(__file__) + '.log',
                     format="{asctime} [{levelname:8}] {process} {thread} {module}: {message}",
                     style="{")
 
 
 async def main():
-    pprint(ubldc.cluster.submit_license(api_secret=api_secret, license_token=license_token), indent=4)
+    pprint(ubldc.cluster.get_cluster_info(), depth=4)
+    await asyncio.sleep(1)
 
 try:
     with BinanceLocalDepthCacheManager(exchange=exchange, ubdcc_address=ubdcc_address) as ubldc: