+++ examples/unicorn_depthcache_cluster_for_binance/ubdcc-stop-all-depthcaches.py
@@ -2,25 +2,34 @@
 # -*- coding: utf-8 -*-
 # ¯\_(ツ)_/¯
 
+from dotenv import load_dotenv
+from unicorn_binance_local_depth_cache import BinanceLocalDepthCacheManager, DepthCacheClusterNotReachableError
 import asyncio
 import logging
 import os
-from pprint import pprint
-from unicorn_binance_local_depth_cache import BinanceLocalDepthCacheManager, DepthCacheClusterNotReachableError
+
+load_dotenv()
 
 exchange: str = "binance.com-futures"
-markets: list = ['1000SHIBUSDT', 'BTCUSDT', 'ETHUSDT']
-ubdcc_address = ""
+ubdcc_address: str = os.getenv('UBDCC_ADDRESS')
 
 logging.getLogger("unicorn_binance_local_depth_cache")
-logging.basicConfig(level=logging.INFO,
+logging.basicConfig(level=logging.ERROR,
                     filename=os.path.basename(__file__) + '.log',
                     format="{asctime} [{levelname:8}] {process} {thread} {module}: {message}",
                     style="{")
 
 
 async def main():
-    pprint(ubldc.cluster.get_cluster_info(), depth=4)
+    dc = ubldc.cluster.get_depthcache_list()
+    for dcl_exchange in dc['depthcache_list']:
+        print(f"Stopping {len(dc['depthcache_list'][dcl_exchange])} DepthCaches for exchange '{dcl_exchange}' on UBDCC "
+              f"'{ubdcc_address}'!")
+        loop = 1
+        for market in dc['depthcache_list'][dcl_exchange]:
+            print(f"Stopping DepthCache #{loop}: {market}")
+            ubldc.cluster.stop_depthcache(exchange=dcl_exchange, market=market)
+            loop += 1
     await asyncio.sleep(1)
 
 try: