+++ CHANGELOG.md
@@ -9,7 +9,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 [How to upgrade to the latest version!](https://unicorn-binance-local-depth-cache.docs.lucit.tech/readme.html#installation-and-upgrade)
 
-## 2.1.1.dev (development stage/unreleased/unstable)
+## 2.2.0.dev (development stage/unreleased/unstable)
+
+## 2.2.0
+### Added
+- ***Experimental*** Client functions for interacting with the UNICORN DepthCache Cluster for Binance 
+
 
 ## 2.1.1
 ### Fixed